feat: refatoração do código
