+++ carteiro-utils.js
