+++ dialogs/finish-talking.js
@@ -1,0 +1,64 @@
+const builder = require('botbuilder');
+
+//Após a exibição do resultado da busca, devemos perguntar o que ele achou do nosso serviço
+module.exports =  [
+    function (session) {
+        builder.Prompts.text(session, `${session.userData.userName}, fui util à você ?`);
+    },
+    function (session, results) {
+        const _msg = results.response;
+
+        session.sendTyping();
+
+        // Creating the Cognitive Services credentials
+        // This requires a key corresponding to the service being used (i.e. text-analytics, etc)
+        if (!process.env.TextAnalyticKey) {
+
+            session.endConversation('Obrigado pela resposta ;)');
+            return;
+        }
+        let _credentials = new CognitiveServicesCredentials(process.env.TextAnalyticKey);
+
+        //Fazer requisição da análise de sentimento da opnião do serviço
+        let _client = new TextAnalyticsAPIClient(_credentials, 'brazilsouth');
+        let _input = {
+            documents: [
+                {
+                    "language": "pt",
+                    "id": "message",
+                    'text': _msg
+                }
+            ]
+        };
+
+        //Validar score da análise e então responder.
+        let _operation = _client.sentiment(_input);
+        _operation.then(function (result) {
+
+            let _responseMessage = '';
+            if (result.errors.length === 0) {
+                const _sentimentScore = result.documents[0].score;
+
+                if (_sentimentScore <= 1 && _sentimentScore > 0.7) {
+                    _responseMessage = "Você é fera!!! Muito obrigado pelo feedback. ;)";
+                }
+                else if (_sentimentScore <= 0.7 && _sentimentScore > 0.4) {
+                    _responseMessage = "Foi muito bom te ouvir, estou sempre melhorando !";
+                }
+                else {
+                    _responseMessage = "Peço desculpa se não fui últil, estarei melhorando minhas buscas ! :'(";
+                }
+            }
+            else {
+                _responseMessage = 'Obrigado pela resposta ;)';
+            }
+
+            session.endConversation(_responseMessage);
+
+        }).catch(function (err) {
+            console.log(err);
+            session.endConversation('Obrigado pela resposta ;)');
+        });
+
+    }
+];
